+++ src/examples/t005.go
@@ -19,9 +19,9 @@ func main() {
 	p.CheckedCmd("set grid y")
 	p.CheckedCmd("set grid z")
 	p.PlotXYZ(
-		[]float{0,1,2,3,4,5,6,7,8,9,10},
-		[]float{0,1,2,3,4,5,6,7,8,9,10},
-		[]float{0,1,2,3,4,5,6,7,8,9,10},
+		[]float64{0,1,2,3,4,5,6,7,8,9,10},
+		[]float64{0,1,2,3,4,5,6,7,8,9,10},
+		[]float64{0,1,2,3,4,5,6,7,8,9,10},
 		"test 3d plot")
 	p.SetLabels("x", "y", "z")
 	p.CheckedCmd("set terminal pdf")
